Moved DiscoveredWearablesAdapter to separate class.
